refactor: better errorMap structure
